reactiverange/backend/requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Cors==4.0.1
SQLAlchemy==2.0.32
python-dotenv==1.0.1
bcrypt==4.1.3
PyJWT==2.9.0
docker==7.1.0
google-generativeai==0.7.2
eventlet==0.36.1
